+++ example/lib/example_route.dart
@@ -42,7 +42,8 @@ FFRouteSettings getRouteSettings({
         arguments: arguments,
         widget: LineBreakingOverflowStyleDemo(),
         routeName: 'LineBreakingOverflowStyle',
-        description: 'workaround for issue 18761. LineBreakingOverflowStyle',
+        description:
+            'make line breaking and overflow style better,workaround for issue 18761.',
       );
     case 'fluttercandies://TextDemo':
       return FFRouteSettings(