+++ example/lib/main.dart
@@ -12,7 +12,6 @@ class MyApp extends StatelessWidget {
   // This widget is the root of your application.
   @override
   Widget build(BuildContext context) {
-
     return OKToast(
         child: MaterialApp(
       title: 'ExtendedText',