+++ lib/src/extended_text.dart
@@ -35,8 +35,10 @@ class ExtendedText extends StatelessWidget {
     this.selectionHeightStyle = ui.BoxHeightStyle.tight,
     this.selectionWidthStyle = ui.BoxWidthStyle.tight,
     this.overflowWidget,
-    this.perfectLineBreakingAndOverflowStyle = false,
+    this.betterLineBreakingAndOverflowStyle = false,
   })  : textSpan = null,
+        // assert(!(betterLineBreakingAndOverflowStyle && selectionEnabled),
+        //    'join zero width space into text, the word will not be a word, the [TextPainter] won\'t work any more.'),
         super(key: key);
 
   /// Creates a text widget with a [InlineSpan].
@@ -64,9 +66,11 @@ class ExtendedText extends StatelessWidget {
     this.selectionHeightStyle = ui.BoxHeightStyle.tight,
     this.selectionWidthStyle = ui.BoxWidthStyle.tight,
     this.overflowWidget,
-    this.perfectLineBreakingAndOverflowStyle = false,
+    this.betterLineBreakingAndOverflowStyle = false,
   })  : data = null,
         specialTextSpanBuilder = null,
+        // assert(!(betterLineBreakingAndOverflowStyle && selectionEnabled),
+        //     'join zero width space into text, the word will not be a word, the [TextPainter] won\'t work any more.'),
         super(key: key);
 
   /// maxheight is equal to textPainter.preferredLineHeight
@@ -209,7 +213,7 @@ class ExtendedText extends StatelessWidget {
   ///
   /// Characters(text).join('\u{200B}')
   ///
-  final bool perfectLineBreakingAndOverflowStyle;
+  final bool betterLineBreakingAndOverflowStyle;
 
 //  ExtendedRenderEditable get _renderEditable =>
 //      _editableTextKey.currentState.renderEditable;
@@ -235,8 +239,12 @@ class ExtendedText extends StatelessWidget {
       children: textSpan != null ? <InlineSpan>[textSpan!] : null,
     );
 
-    if (perfectLineBreakingAndOverflowStyle) {
-      innerTextSpan = joinChar(innerTextSpan, Accumulator(), '\u{200B}');
+    if (betterLineBreakingAndOverflowStyle) {
+      innerTextSpan = joinChar(
+        innerTextSpan,
+        Accumulator(),
+        zeroWidthSpace,
+      );
     }
 
     //_createImageConfiguration(<InlineSpan>[innerTextSpan], context);
@@ -259,12 +267,14 @@ class ExtendedText extends StatelessWidget {
         selectionColor: selectionColor,
         dragStartBehavior: dragStartBehavior,
         onTap: onTap,
-        data: data ?? textSpanToActualText(innerTextSpan),
+        data: (betterLineBreakingAndOverflowStyle ? data?.joinChar() : data) ??
+            textSpanToActualText(innerTextSpan),
         textSelectionControls: selectionControls,
         textHeightBehavior:
             textHeightBehavior ?? defaultTextStyle.textHeightBehavior,
         textWidthBasis: textWidthBasis ?? defaultTextStyle.textWidthBasis,
         overFlowWidget: overflowWidget,
+        strutStyle: strutStyle,
       );
     } else {
       result = ExtendedRichText(
@@ -285,6 +295,7 @@ class ExtendedText extends StatelessWidget {
         textWidthBasis: textWidthBasis ?? defaultTextStyle.textWidthBasis,
         overflowWidget: overflowWidget,
         hasFocus: false,
+        strutStyle: strutStyle,
       );
     }